+++ docs/fr_FR/Changelog.rst
@@ -9,6 +9,7 @@ ChangeLog
 - Interne: eqLogic/configuration: groupEPx => eqModel['variables'].
 - Page EQ/avancé: Ajout d'une section 'variables'. Suppression cas 'telecommande7groups'.
 - Tuya compteur d'énergie: Modèle préliminaire (2691).
+- Page EQuipement: 'variables' migrées sur onglet principal.
 
 240308-BETA-1
 -------------